+++ tests/run-multitests.py
@@ -15,7 +15,13 @@ import itertools
 import subprocess
 import tempfile
 
-run_tests_module = __import__("run-tests")
+from test_utils import (
+    base_path,
+    pyboard,
+    test_instance_epilog,
+    convert_device_shortcut_to_real_device,
+    create_test_report,
+)
 
 test_dir = os.path.abspath(os.path.dirname(__file__))
 
@@ -23,9 +29,6 @@ if os.path.abspath(sys.path[0]) == test_dir:
     # remove the micropython/tests dir from path to avoid
     # accidentally importing tests like micropython/const.py
     sys.path.pop(0)
-
-sys.path.insert(0, test_dir + "/../tools")
-import pyboard
 
 if os.name == "nt":
     CPYTHON3 = os.getenv("MICROPY_CPYTHON3", "python3.exe")
@@ -554,7 +557,7 @@ def main():
     cmd_parser = argparse.ArgumentParser(
         description="Run network tests for MicroPython",
         epilog=(
-            run_tests_module.test_instance_epilog
+            test_instance_epilog
             + "Each instance arg can optionally have custom env provided, eg. <cmd>,ENV=VAR,ENV=VAR...\n"
         ),
         formatter_class=argparse.RawTextHelpFormatter,
@@ -582,7 +585,7 @@ def main():
     cmd_parser.add_argument(
         "-r",
         "--result-dir",
-        default=run_tests_module.base_path("results"),
+        default=base_path("results"),
         help="directory for test results",
     )
     cmd_parser.add_argument("files", nargs="+", help="input test files")
@@ -612,7 +615,7 @@ def main():
             print("unsupported instance string: {}".format(cmd), file=sys.stderr)
             sys.exit(2)
         else:
-            device = run_tests_module.convert_device_shortcut_to_real_device(cmd)
+            device = convert_device_shortcut_to_real_device(cmd)
             instances_test.append(PyInstancePyboard(device))
 
     for _ in range(max_instances - len(instances_test)):
@@ -626,7 +629,7 @@ def main():
                 break
 
             test_results = run_tests(test_files, instances_truth, instances_test_permutation)
-            all_pass &= run_tests_module.create_test_report(cmd_args, test_results)
+            all_pass &= create_test_report(cmd_args, test_results)
 
     finally:
         for i in instances_truth:
