+++ tests/run-natmodtests.py
@@ -9,7 +9,13 @@ import subprocess
 import sys
 import argparse
 
-run_tests_module = __import__("run-tests")
+from test_utils import (
+    base_path,
+    pyboard,
+    test_instance_epilog,
+    get_test_instance,
+    create_test_report,
+)
 
 # Paths for host executables
 CPYTHON3 = os.getenv("MICROPY_CPYTHON3", "python3")
@@ -110,7 +116,7 @@ class TargetPyboard:
             output = self.pyb.exec_(script)
             output = output.replace(b"\r\n", b"\n")
             return output, None
-        except run_tests_module.pyboard.PyboardError as er:
+        except pyboard.PyboardError as er:
             return b"", er
 
 
@@ -221,7 +227,7 @@ def main():
 
     cmd_parser = argparse.ArgumentParser(
         description="Run dynamic-native-module tests under MicroPython",
-        epilog=run_tests_module.test_instance_epilog,
+        epilog=test_instance_epilog,
         formatter_class=argparse.RawDescriptionHelpFormatter,
     )
     cmd_parser.add_argument(
@@ -243,7 +249,7 @@ def main():
     cmd_parser.add_argument(
         "-r",
         "--result-dir",
-        default=run_tests_module.base_path("results"),
+        default=base_path("results"),
         help="directory for test results",
     )
     cmd_parser.add_argument("files", nargs="*", help="input test files")
@@ -257,9 +263,7 @@ def main():
 
     target_truth = TargetSubprocess([CPYTHON3])
 
-    target = run_tests_module.get_test_instance(
-        args.test_instance, args.baudrate, args.user, args.password
-    )
+    target = get_test_instance(args.test_instance, args.baudrate, args.user, args.password)
     if target is None:
         # Use the unix port of MicroPython.
         target = TargetSubprocess([MICROPYTHON])
@@ -283,7 +287,7 @@ def main():
 
     os.makedirs(args.result_dir, exist_ok=True)
     test_results = run_tests(target_truth, target, args, target_arch)
-    res = run_tests_module.create_test_report(args, test_results)
+    res = create_test_report(args, test_results)
 
     target.close()
     target_truth.close()
