+++ tests/run-perfbench.py
@@ -10,9 +10,13 @@ import sys
 import argparse
 from glob import glob
 
-run_tests_module = __import__("run-tests")
-
-prepare_script_for_target = run_tests_module.prepare_script_for_target
+from test_utils import (
+    base_path,
+    pyboard,
+    get_test_instance,
+    prepare_script_for_target,
+    create_test_report,
+)
 
 # Paths for host executables
 if os.name == "nt":
@@ -49,7 +53,7 @@ def run_script_on_target(target, script):
         try:
             target.enter_raw_repl()
             output = target.exec_(script)
-        except run_tests_module.pyboard.PyboardError as er:
+        except pyboard.PyboardError as er:
             err = er
     else:
         # Run local executable
@@ -277,7 +281,7 @@ def main():
     cmd_parser.add_argument(
         "-r",
         "--result-dir",
-        default=run_tests_module.base_path("results"),
+        default=base_path("results"),
         help="directory for test results",
     )
     cmd_parser.add_argument(
@@ -298,9 +302,7 @@ def main():
     M = int(args.M[0])
     n_average = int(args.average)
 
-    target = run_tests_module.get_test_instance(
-        args.test_instance, args.baudrate, args.user, args.password
-    )
+    target = get_test_instance(args.test_instance, args.baudrate, args.user, args.password)
     if target is None:
         # Use the unix port of MicroPython.
         target = [MICROPYTHON, "-X", "emit=" + args.emit]
@@ -328,7 +330,7 @@ def main():
 
     os.makedirs(args.result_dir, exist_ok=True)
     test_results = run_benchmarks(args, target, N, M, n_average, tests)
-    res = run_tests_module.create_test_report(args, test_results)
+    res = create_test_report(args, test_results)
 
     if hasattr(target, "exit_raw_repl"):
         target.exit_raw_repl()
